requirements.txt
streamlit==1.32.0
pymongo==4.5.0
pandas==2.0.3
plotly==5.18.0
python-dotenv==1.0.0
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2